+++ NEWS.md
@@ -4,6 +4,10 @@
 
 * Adds support for new ML methodology implemented in Spark 4.0 (#153). Specific
 ML functions now supported are:
+  * `ft_tokenizer()` *(New for Spark 4.0)*
+  * `ft_stop_words_remover()` *(New for Spark 4.0)*
+  * `ft_hashing_tf()` *(New for Spark 4.0)*
+  * `ft_normalizer()` *(New for Spark 4.0)*
   * `ft_binarizer()` *(New for Spark 4.0)*
   * `ft_bucketizer()` *(New for Spark 4.0)*
   * `ml_linear_regression()` *(New for Spark 4.0)*